Add a poc for checkpoint/restore UDP receive queue
In this commit, the packets in receive queue of restored socket are guaranteed
to have the same data and source IP address as when being checkpointed.

client and server run in 2 different network namespace, server is checkpointed
when a packet has already arrived but not read yet. When restoring a socket is
created in the same local network of server, the dumped packet is then sent to
the server. In order to keep the same source IP address, pktinfo cmsg is used to
change source IP address and socket is set to IP_TRANSPARENT. This can also be
achieve using IP_HDRINCL.